Changes made: Several bug fixes and UI enhancements as follows: 1. Upon changing the dimensions or orientation in the Chart Selector, the default chart subtype was being selected. Now, a check is made to see if the current selected subtype is still supported with the new dimension/orientation. If so, it is maintained. 2. When returning to the Chart Builder from the Chart Selector, the Preview was being displayed even if it was not shown earlier. Fixed this problem by checking if the preview window is to be shown before it is displayed. 3. Added a 'reset()' method that is called when all Marker Lines / Ranges are removed. This method resets the UI to default values that appear before a line / range is added. 4. Added UI to set Grid Column Count (to control how multiple pies are rendered within a single chart 5. Changed the UI for transparency into a slider. Now, the transparency of a color can be in the range from 0 to 255 from the UI to get a translucent color. 6. Cleaned up code in UI classes for unused / unnecessary variables etc.

+++ chart/org.eclipse.birt.chart.ui.extension/src/org/eclipse/birt/chart/ui/swt/composites/FontCanvas.java
@@ -140,7 +140,7 @@ public class FontCanvas extends Canvas implements PaintListener
                 }
                 else if (fdCurrent.getAlignment().getVerticalAlignment().equals(VerticalAlignment.CENTER_LITERAL))
                 {
-                    iStartY = (int) (this.getSize().y / 2);
+                    iStartY = (this.getSize().y / 2);
                     if (bUseSize)
                     {
                         iStartY -= (getStringWidth(gc, sFontName).y / 2);
